Refactor production process details to separate page with routing
