+++ frontend/src/components/Materials/MaterialForm.tsx
@@ -190,7 +190,7 @@ const MaterialForm: React.FC<MaterialFormProps> = ({
         if (mode === 'create') {
           result = await dispatch(createMaterial(materialData)).unwrap();
         } else if (mode === 'edit' && materialId) {
-          result = await dispatch(updateMaterial({ id: materialId, data: materialData as MaterialUpdate })).unwrap();
+          result = await dispatch(updateMaterial({ materialId, materialData: materialData as MaterialUpdate })).unwrap();
         }
 
         setNotification({
